Fix _NET_WM_NAME
Gstreamer does not set the _NET_WM_NAME property on the window it
creates when using `autovideosink` (or ximagesink/xvimagesink for that
matter).

A lot of tools (like Zoom) filter based on this field being non-null, so
the window can't be shared.

This horrible hack finds the first open window that matches our application
name and proceeds to set _NET_WM_NAME on it.

It is very ugly, but it works.