Set _NET_WM_NAME (as server_name) so the window can be shared on screen-sharing software #31
(added fix from  PR #3 from https://github.com/DavidVentura/UxPlay)

Gstreamer does not set the _NET_WM_NAME property on the window it
creates when using autovideosink (or ximagesink/xvimagesink for that
matter).

A lot of tools (like Zoom) filter based on this field being non-null, so
the window can't be shared.

This horrible hack finds the first open window that matches our application
name and proceeds to set _NET_WM_NAME on it.

It is very ugly, and I do not know what I am doing, but it works. Feel free to use this to write something that makes more sense.

---------------------------
changes from PR #31:
use the UxPlay server-name as the widow name.